% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{A_example}
\alias{A_example}
\title{A_example dataset}
\format{
A \code{data.frame} with 10 records. Each row represents one record, with the following columns:
\code{name}, \code{surname}, and \code{city}. Some records can be matched with records in the
\code{B_example} dataset.
}
\usage{
A_example
}
\description{
An example dataset containing artificial personal data.
}
\examples{
data("A_example")
A_example
}
\keyword{datasets}
