% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparators.R
\name{abs_distance}
\alias{abs_distance}
\title{Absolute Distance Comparison Function}
\usage{
abs_distance()
}
\value{
Returns a function taking two arguments, \code{x} and \code{y}, and returning their absolute difference.
}
\description{
Creates a function that calculates the absolute distance between two values.
}
\examples{
cmp <- abs_distance()
cmp(1, 5) # returns 4

}
\author{
Adam Struzik
}
