% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{B_example}
\alias{B_example}
\title{B_example dataset}
\format{
A \code{data.frame} with 12 records. Each row represents one record, with the following columns:
\code{name}, \code{surname}, and \code{city}. Some records can be matched with records in the
\code{A_example} dataset.
}
\usage{
B_example
}
\description{
An example dataset containing artificial personal data.
}
\examples{
data("B_example")
B_example
}
\keyword{datasets}
