% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supervised_learning.R
\name{train_rec_lin}
\alias{train_rec_lin}
\title{Train a record linkage classifier}
\usage{
train_rec_lin(
  A,
  B,
  matches,
  variables,
  comparators = NULL,
  methods = NULL,
  controls_nleqslv = list(),
  controls_kliep = control_kliep()
)
}
\description{
Train a record linkage classifier
}
