% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supervised_learning.R
\name{custom_rec_lin_model}
\alias{custom_rec_lin_model}
\title{Create a custom record linkage model}
\usage{
custom_rec_lin_model(ml_model, vectors)
}
\description{
Create a custom record linkage model
}
