% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparators.R
\name{jarowinkler_complement}
\alias{jarowinkler_complement}
\title{Jaro-Winkler Distance Complement}
\usage{
jarowinkler_complement()
}
\value{
Returns a function taking two string arguments, \code{x} and \code{y},
and returning the complement of the Jaro-Winkler distance.
}
\description{
Creates a function that calculates the complement of the Jaro-Winkler
distance between two strings (i.e., 1 - Jaro-Winkler distance).
}
\author{
Adam Struzik
}
