% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparators.R
\name{jarowinkler_complement}
\alias{jarowinkler_complement}
\title{Jaro-Winkler Distance Complement}
\usage{
jarowinkler_complement()
}
\description{
Jaro-Winkler Distance Complement
}
\author{
Adam Struzik
}
