% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised_learning.R
\name{mec}
\alias{mec}
\title{Unsupervised Maximum Entropy Classifier for Record Linkage}
\usage{
mec(
  A,
  B,
  variables,
  comparators = NULL,
  methods = NULL,
  nonpar_hurdle = FALSE,
  start_params = NULL,
  set_construction = NULL,
  delta = 0.5,
  eps = 0.05,
  controls_nleqslv = list(),
  controls_kliep = control_kliep(),
  true_matches = NULL
)
}
\description{
Unsupervised Maximum Entropy Classifier for Record Linkage
}
\author{
Adam Struzik
}
