% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.rec_lin_model}
\alias{predict.rec_lin_model}
\title{Predict matches based on a given record linkage model}
\usage{
\method{predict}{rec_lin_model}(
  object,
  newdata_A,
  newdata_B,
  set_construction = c("size", "flr"),
  fixed_method = "Newton",
  target_flr = 0.05,
  tol = 10^(-3),
  max_iter = 1000,
  ...
)
}
\description{
Predict matches based on a given record linkage model
}
